+++ CHANGELOG.md
@@ -48,7 +48,7 @@ All notable changes to this project should be documented in this file.
 - A `PatternMatchingMutator` that injects match/case statements to target pattern matching UOPs, by @devdanzin.
 - An `ArithmeticSpamMutator` that injects tight loops with arithmetic operations to target related UOPs, by @devdanzin.
 - A `NewUnpackingMutator` targeting unpacking dicts and single element lists, by @devdanzin.
-
+- A `StringInterpolationMutator` targeting string interpolation UOPs, by @devdanzin.
 
 ### Enhanced